requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
wikipedia
psutil
pyautogui
pyjokes
